Add: Method to find color match if not exact match
After calculating the exact matches where color
and position are correct, I calculate when there
is a correct color match but in the wrong position.

This required removing the potential for double
counting, so calculating the differences between
the code and exact matches and then current
guess with remainder guess was important.