Update: Code to allow repeated colors
The true mastermind game is supposed to allow the
colored pegs to repeat. For example, green should
be able to show up in the code more than once.

Shuffle only provides a unique set of values, so
I used map and sample to get the code to randomly
pick values with the possible repetition of elements.