Update: Test should use random guess
Now I updated the code length in configuration,
the guess should also be randomized (to simulate
a human player) instead of shuffling the code.

This makes it possible to test the feedback method
in the board because the colored matches method
doesn't work correctly if the color pegs are
correct.