Refactor: Make code an instance variable
As I refactored the test for testing board logic,
I realized  I was passing in code as a parameter
in almost every method. I decided to refactor the
Board class to use code as an instance variable
so I didn't have to pass in code so often.

Since the secret code stays the same throughout
an entire game, it makes sense to me for now to
define it as an instance variable.