Add: State-tracking variables and logic to update
I created the game state class and added new
methods to track the state of the game. Most of
the logic comes from the board method.

Even though there's a seeming overlap of methods
between the GameState and Board classes, I'm
violating the DRY principle to favor of SOLID
principles (separation of responsibility).