Refactor: Move methods to private
As I'm calling interface from Game class, I feel
like Game class needs to delegate more logic to
the Interface class. Therefore, Interface class
is responsible for providing a valid guess to the
Game class, how it does so is its own business.

This prompted me to put most of the methods I first
set as public to private so Game class wouldn't
call them to request a guess again and wouldn't
be  responsible for validating inputs.