Fix: Bug where Reset wasn't working
Reset was pasted from the Tic-Tac-Toe game but it
doesn't work. Instead, I pushed the logic to play
another game to the main execution script, which
allows the Game class to fully focus on playing 1
game instead of also managing multiple rounds.

This required deleting the reset method, simplifying
Game and pushing the loop logic to main execution.