Add: Method to sort codes to make new guess
When the feedback adds up to 4, all the colors
are correct but they may be in incorrect positions.

Instead of randomly guessing across all codes, I'd
rather use some kind of random guessing within
the correct colors. Ideally, I'd like to implement
some kind of bubble sort to switch the order of
the characters but I'd need to keep track of
indices, so I haven't implemented that yet.