Update: Sort method to remove irrelevant codes
I have the building blocks to create the logic I'm
looking for. I aim to repurpose existing methods
to detect irrelevant codes when feedback_sum == 4.

I know all the correct colors are in last_guess.
I just need to reorder them and randomly pick a
code with all the colors in last_guess as the
guess.

I removed unnecessary instance variables and used
hash to track the counts of the colors. Then I
compare the counts of colors in last_guess to the
counts in all_codes. I only keep the codes in
all_codes with the same count as the colors in
last_guess.