Fix: Bug with array subtraction removing too much
Ruby array subtraction removed all instances of
the color instead of the first result.

For example, if the secret code is CPPY, array
subtraction removed all 'P' so the resulting array
shows ['C'] when I want to return ['C', 'P'].

It should only subtract the first instance of the
values in the array, not all instances.

With Claude's help, I wrote an explicit method
for performing the array subtraction.