Fix: Bug causing nils to enter into the guesses
When secret code was CPPY, the last code in
possible_codes was deleted so possible_codes was
nil.

This created a guess like this:

[nil, "P", "C", "Y"]

because possible_codes[0] is nil.

By this time, the computer has gone through all
possible guesses of colors. The problem is here:
Guess: PPCY - Hint: Exact matches 2, Correct colors 1

I fixed this by changing the method of how it
calculated updating the array with a check.